+++ src/main/java/com/easyerp/model/dto/EstoqueResponse.java
@@ -1,11 +1,9 @@
 package com.easyerp.model.dto;
 
-import java.math.BigDecimal;
-
 import com.easyerp.domain.entidade.Estoque;
 
-public record EstoqueResponse(BigDecimal estoqueAtual) {
+public record EstoqueResponse(Integer estoqueAtual) {
 	public EstoqueResponse(Estoque estoque) {
-		this(estoque.getQuantidade());
+		this(estoque.getQuantidade().intValue());
 	}
 }
